fix(pairing): update extension import paths after src->extensions refactor
