+++ extensions/telegram/src/dm-access.ts
@@ -1,10 +1,13 @@
 import type { Message } from "@grammyjs/types";
 import type { Bot } from "grammy";
-import type { DmPolicy } from "../config/types.js";
-import { logVerbose } from "../globals.js";
-import { issuePairingChallenge } from "../pairing/pairing-challenge.js";
-import { buildPairingReply, type PairingMessageConfig } from "../pairing/pairing-messages.js";
-import { upsertChannelPairingRequest } from "../pairing/pairing-store.js";
+import type { DmPolicy } from "../../../src/config/types.js";
+import { logVerbose } from "../../../src/globals.js";
+import { issuePairingChallenge } from "../../../src/pairing/pairing-challenge.js";
+import {
+  buildPairingReply,
+  type PairingMessageConfig,
+} from "../../../src/pairing/pairing-messages.js";
+import { upsertChannelPairingRequest } from "../../../src/pairing/pairing-store.js";
 import { withTelegramApiErrorLogging } from "./api-logging.js";
 import { resolveSenderAllowMatch, type NormalizedAllowFrom } from "./bot-access.js";
