+++ src/config/types.imessage.ts
@@ -5,7 +5,6 @@ import type {
   GroupPolicy,
   MarkdownConfig,
 } from "./types.base.js";
-import type { PairingMessageConfig } from "../pairing/pairing-messages.js";
 import type {
   ChannelHealthMonitorConfig,
   ChannelHeartbeatVisibilityConfig,
